+++ src/OBD/data/BitEncoded/Oxygen/BankOxygenSensors4BankCollection.cpp
@@ -95,7 +95,7 @@ DataObjectStateCollection BankOxygenSensors4BankCollection::setValueFromString(s
     }
 
     int i;
-    for (i = 0; i < allSensors.size(); i++) {
+    for (i = 0; i < (int) allSensors.size(); i++) {
         DataObjectStateFactory::merge(rs, allSensors.at(i)->setValueFromString(parts.at(i)));
     }
